improve jds modularization
